Simplify internal liquid optics calls
Simplify call tree for liquid optics. Previously, there was an extra
layer of subroutines that looped over columns and levels and called a
separate routine to interpolate for one point. Instead, just put the
interpolation into the loop, as was already done for the ice optics. The
ice and liquid optics now have a similar structure, and it is a little
easier to figure out what is going on.